Oracle client for Docker container
Why these changes are being introduced:

This application requires that Oracle client libraries are available for the
python oracledb library to use.

This does not include installation instructions for local development in arm64
which is possible and forthcoming.

How this addresses that need:
* linux x86 and mac arm64 Oracle 19.x client zip files added to /vendor folder
* Dockerfile updated to unpack libraries to /opt/lib for use in python context

Side effects of this change:
* Docker container has Oracle client libraries available

Relevant ticket(s):
* None