Setup Config class and luigi config file
Why these changes are being introduced:

Luigi supports a custom configuration file.  Given we want to run this app primarily
from the hrqb folder, this sets up an env var pattern to define where the config file is.

This also implements a similar design to other apps of a 'Config' class with
dot notation available env vars.

How this addresses that need:
* Creates Config class for app to use
* Sets required env var LUIGI_CONFIG_PATH in Config class

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-10