Refactor base Tasks
Why these changes are being introduced:

While working on building out actual chains of Tasks in Pipelines,
it was determined that it would be beneficial to programmatically
set the Target output paths for Tasks.  To support this, a pipeline name
will get passed throughout that help name these artifacts.  This will
allow a single Task to get reused in different pipelines, where the output
Target will have a filename unique the pipeline.

With these changes, it was also determined that testing could be
simplified to use some Tasks as fixtures.

How this addresses that need:
* HRQBTask.path is a dynamic property that constructs a path based
on a configurable targets directory, pipeline name, task stage, and
task name
* Task fixtures were created for extract, transform, load, and pipeline
* PandasPickleTarget is limited to only returning DataFrames as it was
determined that Series are likely not going to get used

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-13