Method to init HRQBPipelineTask from classpath
Why these changes are being introduced:

The CLI command group is expecting a pipeline name and optional
pipeline module, both as strings, to find a pipeline task to run.
Because these are strings, it requires loading that actual python
class via imports and dynamic loading.  While functional before,
it was a little opaque to have that complex logic in the CLI, where
it would be better suited as a dedicated method on the HRQBPipelineTask
class which is the target parent class for anything imported.

How this addresses that need:
* Adds method HRQBPipelineTask.init_task_from_class_path dedicated
to finding an instance via module and class strings to import and
instantiate

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-13