Add DWClient for Data Warehouse work
Why these changes are being introduced:

A primary source of data for this application to work with
is data coming from the MIT Oracle Data Warehouse.  A normalized
way of connecting and performing queries is needed.

How this addresses that need:
* Creates new DWClient to connect to, and query from, a remote
database.
* DWClient includes a method execute_query that uses pandas
to return a DataFrame of the results, leaning into the
DataFrame-first nature of the PandasPickleTasks.

Side effects of this change:
* Connection possible to external Data Warehouse

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-11