Convert convenience method target to property
Why these changes are being introduced:

In this project, it was decided to include a convenience
attribute 'target' that was a method, to better indicate
it is returning a luigi.Target instance.  Default luigi Tasks
require an output() method that returns a Target.

Converting this to a property further leans into this
project-specific convention of providing a convenience
attribute, making it an even easier to use property.

How this addresses that need:
* Converts HRQBTask.target() into a property
* defines HRQBTask.output() to avoid defining this
in classes that extend this opinionated project class

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-13