Add record count on base targets
Why these changes are being introduced:

Debugging error will be easier if counts are available for
the total number of records processed by a Task.  This will
reveal tasks that are not returning the expected number of
records for downstream tasks.

How this addresses that need:
* Adds new properties on base Targets to produce a record
count by reading the output Target data

Side effects of this change:
* None