Add ECR commands/makefile
Add the ecr and makefile commands for ECR interaction with AWS.
As a part of our normal operation, we add an integration with aws in an action that publishes images to the ecr repo in the dev1 aws account.  We also add makefile commands for manually publishing the images if needed.