Env var for parallel luigi workers
Why these changes are being introduced:

When a pipeline is run via the luigi runner, it can support multiple parallel
workers.  It is unlikely we will need more than one concurrent worker, which
means we do not need any threading or multiprocessing overhead.  But this
env var allows for that if needed.

How this addresses that need:
* Add new optional env var LUIGI_NUM_WORKERS
  * If not set, luigi.build() is run with 1 worker

Side effects of this change:
* By default, pipelines are run with a single worker

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-17