Remove data for specific tasks
Why these changes are being introduced:

It is helpful when developing on a task to have the ability to programatically
remove the output target data from only that task.  This should also apply when
including the --cleanup flag for a pipeline run, if a single task is specified,
where only the target data for that task would be removed.

How this addresses that need:
* Normalize --task argument on remove-data and run CLI commands
* Updates docstrings and README to match

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-17