Move CLI target task arg to pipeline group command
Why these changes are being introduced:

When running sub-commands for the pipeline command, e.g.
remove-data or run, you can specifiy a target task to focus on
from the pipeline.  Formerly this was a CLI arg on the sub-commands,
but this could be hoisted into the parent pipeline command and normalize
behavior.

How this addresses that need:
* Moves --task CLI arg to pipeline group command

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-17