Remove files to group tasks by type
Why these changes are being introduced:

Getting into the task work, it became evident that it would be
preferable to group all tasks related to a single table, in a
single file.  This does not prevent importing tasks as dependencies
from other files, but it groups like tasks.

How this addresses that need:
* Removes tasks/extract.py, tasks/transform.py, tasks/load.py

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-21