Tests for Employees tasks
Why these changes are being introduced:

Unit testing will be somewhat minimal for actual tasks, given much logic
is pushed down to base tasks.  But some low hanging fruit is valuable,
like if the tasks have SQL files to read, if explicit transformations are as expected, etc.

This sets a minimal baseline to aim for for future tasks.

How this addresses that need:
* Added tests/tasks directory with a one-to-one for task files
* Add preliminary unit tests for Employees table tasks

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-21