Tasks for LibHREmployeeAppointments table
Why these changes are being introduced:

HRQBClient needs to support a unique pipeline that will load static
data from HR into a Quickbase table.  This pipeline will be run
quite rarely, and would be suitable to do so manually from the CLI
locally; an initial load and potentially bulk updates from HR.
After that, this table will be managed directly in QB.  The data
in this table is utilized by other tasks and pipelines.

How this addresses that need:
* Creates new UpdateLibHRData pipeline and associated tasks

Side effects of this change:
* Ability to load and update LibHREmployeeAppointments table via
HRQBClient

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-15