Limit Employees data to 2019 or newer
Why these changes are being introduced:

One requirement of the Quickbase implementation is data going
back to 2019.  Until now, no data limits were placed on data warehouse
queries.  This begins to establish this for pre-existing tasks, and will
be enforced for future ones as well.

How this addresses that need:
* Limits Employee records to those with Employee Appointments that
end on or after 2019

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-22