ETL tasks for Employee Leave
Why these changes are being introduced:

Another primary set of tables that is loaded from data warehouse data is
around employee leave (absences).

Noting that it is still under consideration how seniority accrual is
determined for a given absence.  Placeholders for this field are in place
and documented as a WIP.

How this addresses that need:
* Creates ETL tasks for 'Employee Leave' and 'Employee Leave Types' tables
* Continues to refactor shared tasks to hrqb.tasks.shared as needed
* Stub field 'Accrue Seniority' until final logic for calculating this is
determined

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-28