Optional field return on QB upsert
Why these changes are being introduced:

It was determined that large upserts could result in memory pressure
due to the response from quickbase which returned every record
and field value.  The memory footprint for upserting is reasonble,
but this JSON response was parsed into memory which is not memory
efficient.

Ultimately, the modified record field values are not needed most
of the time, so defaults to not returning.  But if the flag
is set to True, field values can be returned on upsert if needed.

How this addresses that need:
* Add return_fields parameter to preparing upsert payload

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-32