Mint MD5 hash values for Key field
Why these changes are being introduced:

It was determined that columns used from the data warehouse, expected to be
enduringly the same and unique, where changing across warehouse runs.  These
values were used for Quickbase merge fields, but this would not work if the
values changed.

A value was needed that uniquely identifies the row for a handful of tables,
across warehouse runs over time.

How this addresses that need:
* Updates multiple tables to mint an MD5 value from an ordered list of
field values that are known to uniquely identify a row

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-36