QBClient able to delete records
Why these changes are being introduced:

For development, it is sometimes required that some or all records from a QB
table are deleted.  This is strangely not trivial in QB.  Paginated deletes in
the GUI are slow and time consuming, and the API requires a somewhat cryptic
query payload.  A testable, normalized way of deleting records is helpful.

How this addresses that need:
* QBClient gets two new methods, delete_records() and
delete_all_table_records()

Side effects of this change:
* None

Relevant ticket(s):
* None