Add reasons to salary history Key field
Why these changes are being introduced:

After possible updates to data warehouse data coming in,
it was determined that additional fields were needed to
mint a unique merge field 'Key' for the Employee Salary
History table.  By adding the HR transaction reason and
type, this is enough to gaurantee the merge field is
unique.

How this addresses that need:
* Adds 'Related Salary Change Type' and 'Salary Change Reason'
as fields for minting unique key

Side effects of this change:
* All records were truncated from Quickbase and recreated
as this change effects the unique merge field

Relevant ticket(s):
* None