Add integrity check pattern
Why these changes are being introduced:

Unit tests are valuable for checking that tasks perform as expected given
inputs and outputs, but they do not help with asserting actual data from
the data warehouse is handled correctly.  To this end, run time 'integrity
checks' are helpful to check for edge cases and bad data before a task is
considered successful.

How this addresses that need:
* New decorator HRQBTask.integrity_check that can be applied to methods
in any task class
* New IntegrityCheckError exception type
* All registered integrity checks are run before a task is considered
successful, throwing an IntegrityCheckError if they are not

Side effects of this change:
* In normal operations, a task will fail and thus downstream tasks
will not fire, if a task fails integrity checks

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-27