Merge QB employee appointments via Key field
Why these changes are being introduced:

After the removal of the volatile 'HR Appointment Key' from a couple
of QB tables, this broke merges in transform tasks for employee leaves
and salary history.  Now, instead of merging on this volatile field,
we need to merge on the 'Key' field which is enduring based on an MD5
hash of employee appointment data.

How this addresses that need:
* TransformEmployeeLeave and TransformEmployeeSalaryHistory both merge
on QB Employee Appointments data by calculating a matching 'Key' value
from data warehouse data

Side effects of this change:
* Rows in Quickbase will not appear modify and leaves and salaries
will not drift away from appointments

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-40