Send sentry message on upsert errors
Why these changes are being introduced:

When performing upserts to Quickbase, even if some records have
errors, the HTTP response will be 200.  It is required to parse
the response and see what specific records had errors.  We capture
this grain in a method to aggregate upsert results across all tasks,
but we only logged it.

Now that all upsert errors have been resolved as a baseline, anything
new indicates new data or an edge case we have not encountered.  As
such, it is a good time to introduce more warning when this happens.

How this addresses that need:
* Manually send a Sentry message when upsert errors detected

Side effects of this change:
* Sentry messages sent

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-46