Add effective salary change percent value
Why these changes are being introduced:

Formerly we calculated percentages for the base salary
changing from the previous row, and the temporary salary
changing the base of the current row.

Now, to paint a fuller picture, we are calculating an
'Effective Salary Change' percent.  This takes the effective
salary from the current row and calculates a percent change
from the effective salary of the previous row.

This way, when a temporary base salary ends, we might see a
negative number for this 'Effective Salary Change' percent, as
the effective salary has decreased.

How this addresses that need:
* Calculates new 'effective_change_percent' field in task 'TransformEmployeeSalaryHistory'
  * Quickbase updated with new field that this writes to
* Added test that confirms calculated effective salary change is expected

Side effects of this change:
* New data written to 'Employee Salary Change' table in Quickbase

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-52