Updates to termination reason and date
Why these changes are being introduced:

Additional feedback from HR has shown where the termination
reason and dates could be improved.

There are some edge cases that make this kind of tricky, e.g.
people that have appointments terminated, but get rehired.  Or,
employees that transfer to another DLC in the organization, where
there is no termination reason in the data warehouse, but the end
date should be their last library appointment.

How this addresses that need:
* Reworks CTEs in the employees query to focus on the
last or current library appointment for an employee
* This is used in the main body of the query to fill end date
and termination reason

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/HRQB-20