Remove end date as part of appointment unique key
Why these changes are being introduced:

An employee appointment's end date was used as part of its
enduring unique key in Quickbase, but this was flawed.  When
an appointment ended, the end date was changed from 2999-12-31 to
the actual end date, changing the unique key, and effectively
leaving an extra row in Quickbase.

How this addresses that need:

Removing the end date as part of the employee appointment key is
sufficient to make them enduringly unique.

Side effects of this change:
* This required dropping and recreating the Employee Appointments
table via a manual run, but otherwise no changes.

Relevant ticket(s):
* None