Update dependencies, pin luigi
Why these changes are being introduced:

A dependencies update was long overdue but luigui 3.6 brings
a considerable amount of linting errors that will need addressing
(or blanket ignoring).  Currently, this project works well with
luigi 3.5.1 so felt safe to pin this for the time being
and allow other updates to proceed.

How this addresses that need:
* updates dependencies
* pins luigi to 3.5.1

Side effects of this change:
* luigi not automatically updated, but will get revisited
during next maintenance

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1194