Remove LibHR table management
Why these changes are being introduced:

For reasons beyond the scope of this commit, it has been determined that
having a table like "LibHR Employee Appointments" (LIBHR) distinct from the
more widely used "Employee Appointments" (EA) is more problematic than
it is helpful.  As such, it has been decided that data formerly found
in the LIBHR table will now be directly added and managed in the EA
table.

How this addresses that need:
* Completely removes any tasks associated with the LIBHR table
* Removes any copying of data to the EA table from the LIBHR table

Side effects of this change:
* LIBHR table is no longer a concern for this middleware

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1194