Unpin luigi and skip type checking broadly
Why these changes are being introduced:

luigi=3.5.2 introduced a litany of type checking errors
for our usage.  As such, we had luigi pinned to 3.5.1
for some time.  With the introduction of pip-audit, and
a found vulnerability for 3.5.1, and bumping python to
3.12, it was time to finally address this.

We fairly heavily extend luigi.Task and luigi.LocalTarget,
and use those base extensions throughoug the project.
Because luigi is poorly typed, even with the `luigi.mypy` module,
we had cascading type check errors.

Manually creating .pyi stubs was explored -- and it works --
but the benefit did not seem to outweigh the cost of creating
and maintaing them.  Particularly since most of this project
is luigi task boilerplate; each task, e.g. ExtractDWEmployees,
is basically just class definitions with a minimal amount of
actual business logic.  This was the appeal of luigi, and it's
why these .pyi files would be nearly doubling the codebase.

The ultimate solution feels to just skip type checking
for luigi based classes, given a) luigi is poorly typed
already, and b) the signatures are effectively the same for
each task.  IDE's will still benefit from understanding
what classes can do, but we skip linting errors.

How this addresses that need:
* Unpins luigi, resulting in install of luigi=3.6.0 at time
of this commit
* Adds explicit type checking ignores for hrqb.base.* and
hrqb.task.*
* Remove vulnerability ignore now no longer present in luigi

Side effects of this change:
* Luigi is updated, but type checking is lost for most
luigi tasks (at least temporarily)

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1258