Add integrity checks for row counts
Why these changes are being introduced:

For some tasks, e.g. transform tasks for 'Employee Appointments' and
'Employee Salary History', we know in advance that the number of rows we are transforming
and preparing for upsert should always exceed the number of rows currently in Quickbase
for that table.  This is because upserts for these tables are all-time data.  If there are
more rows in Quickbase than the upsert, this indicates a problem, usually with how we
are constructing the unique merge field.

For example, we had a situation where 'Employee Appointment' had two rows in Quickbase
that should not have been there, due to a bug where an employee leaves the library but
goes to another MIT position, but we still pickup the new appointment as a library a
appointment.  This integrity check would flag would have flagged the issue for analysis
versus stumbling on it in the data.

How this addresses that need:
* Adds new integrity checks

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1439