Remove end dates from Employee Salary History merge values
Why these changes are being introduced:

Two end dates were used in the construction of a unique, enduring
merge field value for the Employee Salary History transform step:
the associated appointment end date and the end date of the salary
adjustment.

However, it's quite normal for these values to change over time, e.g.
when apointments end or salary adjustments end.  We similarly removed
an appointment's end date from the Employee Appointments table's merge field
construction table for this same reason.

The result is "stale" rows in the Employee Salary History table.

How this addresses that need:

By removing the appointment and salary adjustment end dates, we construct
a hash value that serves as a unique and enduring value for that salary
history row.  In the event an appointment changes, or a new salary history
row is added (e.g. yearly review increase), these values will remain the same
for pre-existing rows and update them versus adding new rows.

Again, this was a fix take in the Employee Appointments table at one point,
but it was missed we needed to do this in Employee Salary History.

Side effects of this change:
* No more "stale" rows in the Employee Salary History table

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1439