Refine logic for Employee Salary History unique key
Why these changes are being introduced:

A recently added integrity alert triggered for the 'Employee Salary
History' table, indicating we had a row in Quickbase that no longer
aligned with data warehouse data.  As has happened before, it appears to
be shifting data warehouse values, this time for "End Date" and
reason and type data, that modifies the unique key.  When the unique key
changes, we get stale data in Quickbase because we are adding rows instead
of updating them.

How this addresses that need:

After some analysis, and discussion with HR, it was determined we could remove
some fields from the unique key logic which keeps it consistent between
data warehouse changes.  Part of this work was also skipping any salary
history with a reason of "Lump Sum" which has only showed up once, and is the
only time we might get multiple values for a single person, on a single day,
for a single appointment.  Similarly, we already skip "Supplemental Pay" for
this reason.

Side effects of this change:
* Improve merge field values, reduce stale data in Salary History table.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1450