Job title links to Pay Grade table
Why these changes are being introduced:

When the Pay Grade table was introduced, the Job Titles table now
has a pointer to a pay grade.  When the daily middleware runs, we
need to provide a Pay Grade record id instead of the pay grade
literal.

How this addresses that need:

For the Job Title task, retrieve all pay grades.  If the pay grade
is found, use the pay grade record identifier, else set to None.

Side effects of this change:
* No more warning during daily runs.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1534